+++ components/src/DropdownMenu/index.js
@@ -1,3 +1,3 @@
-export { default as Dropdown } from "./Dropdown";
+export { default as DropdownMenu } from "./DropdownMenu";
 export { default as DropdownButton } from "./DropdownButton";
 export { default as DropdownLink } from "./DropdownLink";